Split generic call annotation diagnostics
