+++ tests/generic_diagnostics.rs
@@ -32,6 +32,8 @@ mod annotations;
 mod behavior_impls;
 #[path = "generic_diagnostics/bounds.rs"]
 mod bounds;
+#[path = "generic_diagnostics/call_site_annotations.rs"]
+mod call_site_annotations;
 #[path = "generic_diagnostics/composite_annotations.rs"]
 mod composite_annotations;
 #[path = "generic_diagnostics/constructors.rs"]
